analysis/turaco/requirements.txt
lark-parser==0.12.0
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
